Move project to qmake
