+++ include/Ssh2Process.h
@@ -21,19 +21,54 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-#include "Ssh2Debug.h"
-#include "Ssh2Types.h"
+#pragma once
 
-#include <QtCore/QDebug>
+#include <QIODevice>
 
-using namespace qlibssh2;
+#include "Ssh2Channel.h"
 
-bool qlibssh2::checkSsh2Error(const std::error_code& error_code)
+#include "QLibSSH2_global.h"
+
+
+namespace qlibssh2
 {
-    return error_code == ssh2_success || error_code == Ssh2Error::TryAgain;
-}
 
-void qlibssh2::debugSsh2Error(const int ssh2_method_result)
+class QLIBSSH2_EXPORT Ssh2Process : public Ssh2Channel
 {
-    qDebug() << "Ssh2 error: " << ssh2_method_result;
-}
+    Q_OBJECT
+    Q_ENUMS(ProcessStates)
+    Q_PROPERTY(ProcessStates processState READ processState NOTIFY processStateChanged)
+public:
+    enum ProcessStates {
+        NotStarted,
+        Starting,
+        Started,
+        FailedToStart,
+        Finishing,
+        Finished
+    };
+
+    ProcessStates processState() const;
+    void checkIncomingData() override;
+
+signals:
+    void processStateChanged(ProcessStates processState);
+
+protected:
+    Ssh2Process(const QString& command,
+                Ssh2Client* ssh2_client);
+
+private slots:
+    void onSsh2ChannelStateChanged(const ChannelStates& state);
+
+private:
+    void setSsh2ProcessState(ProcessStates ssh2_process_state);
+    std::error_code execSsh2Process();
+
+    const QString command_;
+    ProcessStates ssh2_process_state_;
+
+    friend class Ssh2Client;
+};
+
+} // namespace qlibssh2
