--- test_package/example.cpp
@@ -1,8 +1,0 @@
-#include "QLibssh2/Ssh2Client.h"
-
-using namespace qlibssh2;
-
-int main()
-{
-    Ssh2Client client{{}};
-}
